% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_frequency.R
\name{compute_frequency}
\alias{compute_frequency}
\title{Calculate frequency distributions}
\usage{
compute_frequency(
  x,
  group = NULL,
  round.percent = 0,
  group.exclude.levels = NULL
)
}
\arguments{
\item{x}{a vector}

\item{group}{optional column to stratify frequency counts by}

\item{round.percent}{integer count of places to include in percentage}

\item{group.exclude.levels}{optional vector with levels of group to exclude}
}
\value{
a dataframe with the levels of stratified
}
\description{
Returns a dataframe with columns of frequency, percent, and ratio
}
\examples{
\dontrun{
# create some mock patient data
size = 2500
unif1 = runif(size)
unif2 = runif(size)
patient_id = sample(1:1000000,size)
mock_data = as.data.frame(patient_id) \%>\%
  dplyr::mutate(binary_var1 = ifelse(unif1<0.8,1,0),
                rare_event =ifelse(unif2>0.95,1,0))
# get distribution of rare_event in the entire dataset
compute_frequency(mock_data$rare_event)
# get distribution of rare_event stratified by binary_var1
compute_frequency(mock_data$rare_event, mock_data$rare_event,group.display.levels = c(1))
}

}
