% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_numuniv.R
\name{compute_numuniv}
\alias{compute_numuniv}
\title{Calculate numeric univariate summary statistics}
\usage{
compute_numuniv(x, group = NULL, round.places = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{group}{optional character string indicating the column to stratify by for summary statistics}

\item{round.places}{integer count of places to include for each summary statistic}
}
\value{
a dataframe with columns corresponding to each summary statistic and row-wise observations for each level of the group variable, if applicable
}
\description{
The following univariate summary statistics are reported for a numeric vector:
mean, standard deviation, minimum, maximum,
median, 25th percentile, 75th percentile, percent and count of all missing and non-missing values,
 Missing values are removed for calculation of summary statistics
}
