% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalues.R
\name{format_pvalues}
\alias{format_pvalues}
\title{Round and format p-values}
\usage{
format_pvalues(x, method = c(1, 2), lead.zero = TRUE)
}
\arguments{
\item{x}{numeric p-value input}

\item{method}{integer corresponding to desired rounding convention. See Details.}

\item{lead.zero}{if FALSE, no 0 will be reported in the place before the decimal. Defaults to TRUE.}
}
\value{
a character formatted per the specified method
}
\description{
Converts numeric p-values to rounded and formatted characters according to common rounding conventions.
}
