% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_frequency.R
\name{report_frequency}
\alias{report_frequency}
\title{Report formatted frequency distributions}
\usage{
report_frequency(
  d,
  cols,
  group = NULL,
  round.percent = 0,
  format = TRUE,
  group.exclude.levels = NULL,
  col.exclude.levels = NULL
)
}
\arguments{
\item{d}{a dataframe}

\item{cols}{a vector of column names for which to report the frequency distribution}

\item{group}{optional column name to stratify by}

\item{round.percent}{integer count of decimal places to include in percentage reporting}

\item{format}{if TRUE (default) then a formatted character column is generated with both the count and percentage concatenated as: "count (percentage)"}

\item{group.exclude.levels}{optional vector with levels of group to exclude from report}

\item{col.exclude.levels}{optional vector with levels of variables in cols to exclude from report}
}
\value{
a dataframe with the levels of stratified
}
\description{
Returns a dataframe with columns of frequency, percent, and ratio
}
\examples{
\dontrun{
# some mock patient data
size = 2500
unif1 = runif(size)
unif2 = runif(size)
patient_id = sample(1:1000000,size)
mock_data = as.data.frame(patient_id) \%>\%
 dplyr::mutate(binary_var1 = ifelse(unif1<0.8,1,0),
               binary_var2 = ifelse(unif1<0.4,1,0),
                rare_event =ifelse(unif2>0.95,1,0))
# summarize patients for binary_var1 and cat_var1
report_frequency(mock_data, cols= c("binary_var1","cat_var1"))
# stratify summary by outcome rare_event, omitting patients negative (0) for binary_var1
report_frequency(mock_data, cols= c("binary_var1","cat_var1"),group = "rare_event",col.exclude.levels = 0)
}

}
