% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_unadjusted.R
\name{report_unadjusted}
\alias{report_unadjusted}
\alias{Passed}
\alias{on}
\alias{to}
\alias{the}
\alias{broom::tidy()}
\alias{functions}
\title{Report bivariate effect size}
\usage{
report_unadjusted(
  d,
  model.method = c("glm_binomial", "glm_gaussian", "survival_coxph"),
  outcome.var,
  time.var,
  model.vars,
  round.estimate = 2,
  p.round.method = 1,
  p.lead.zero = TRUE,
  ...
)
}
\arguments{
\item{d}{A dataframe}

\item{model.method}{String containing target model. Options are "glm_binomial", "glm_gaussian", and "survival_coxph".}

\item{outcome.var}{String containing the name of the outcome variable present in d.}

\item{time.var}{String containing the name of the time variable present in d. Must be specified if "survival_coxph" is selected.}

\item{model.vars}{Vector of strings containing the names of independent model variables for which bivariate analysis will be performed.}

\item{round.estimate}{An integer representing the number of places to which the model estimate and confidence intervals are rounded.}

\item{p.round.method}{Integer corresponding to desired rounding convention. See Details.}

\item{p.lead.zero}{Boolean: if FALSE, no 0 will be reported in the place before the decimal. Defaults to TRUE.}

\item{report.inverse}{A vector of strings containing the names of independent model variables for which the inverse of the estimate is desired for reporting.}

\item{verbose}{Boolean: if TRUE (Default), see messages about model outputs warning of possible common errors.}
}
\value{
A dataframe with an labeled model summary
}
\description{
Bivariate modeling is performed independently for each variable in a vector of supplied predictors with options to format reported output.
}
\details{
# p.round.method
Two methods are currently defined. Select 1 (default) to round values above 0.10 to two digits. Select 2 to round values above 0.10 to 1 digit.
APA format is equivalent to method=1 when lead.zero=FALSE.
}
