% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwekit-package.R
\docType{package}
\name{rwekit-package}
\alias{rwekit}
\alias{rwekit-package}
\title{rwekit: Simple reporting toolkit for real-world evidence analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simplifies reporting of common analyses in real-world evidence studies with functions for Table 1 customization and model object annotation.
}
\author{
\strong{Maintainer}: Lindsey Kornowske \email{lindsey.kornowske@yahoo.com}

}
\keyword{internal}
